rename email settings dir and pages to communications
