+++ packages/twenty-front/src/pages/settings/communications/assets/cover-dark.png
